rag_assessment/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "rag-assessment"
version = "1.0.0"
description = "Context-Aware Retrieval Engine: Semantic RAG & Vector Search"
requires-python = ">=3.10"
dependencies = [
    "sentence-transformers>=3.3",
    "faiss-cpu>=1.9",
    "numpy>=1.26",
    "rich>=13.9",
    "streamlit>=1.41",
    "plotly>=5.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-mock>=3.14",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
